% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatOdyData.R
\name{CalcSalinity}
\alias{CalcSalinity}
\title{Calculate salinity from SpCond}
\usage{
CalcSalinity(spcond)
}
\arguments{
\item{spcond}{specific conductance}
}
\description{
generates salinity vector from specific conductance vector
}
\examples{
CalcSalinity(50)
}
