% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateSimpleLevels.R
\name{calculate_levels_simple}
\alias{calculate_levels_simple}
\title{Calculate compensated water levels above sensor using a pair of baro pressure and water pressure files}
\usage{
calculate_levels_simple(wll_csv, baro_csv, wll.type, baro.type = solinst,
  output.dir, water.density, baro.df = NULL, sensor.elev, navd.dir)
}
\arguments{
\item{wll_csv}{the path to a csv file exported directly from the logger software}

\item{baro_csv}{the path to a csv file exported directly from the baro logger software}

\item{wll.type}{the brand of water logger: either 'Hobo' or 'Solinst'}

\item{baro.type}{the brand of baro logger: either 'Hobo' or 'Solinst' or 'df' if using a data frame}

\item{output.dir}{the path to the output directory for the saved files}

\item{water.density}{the average water density for the site, in g/L}

\item{baro.df}{a data frame of baro pressure data and time if not using a CSV. Must have fields "time" and "baro.pressure"}

\item{sensor.elev}{an optional field of absolute sensor elevation}

\item{navd.dir}{the path in which to save logger data with NAVD elevations}
}
\description{
This function performs barometric compensation and density compensation and save the files
}
\examples{
calculate_levels_simple(wll_csv = wl.files[[1]],
baro_csv = 'my_baro.csv',
wll.type = 'Solinst',
baro.type = "Solinst",
output.dir = 'CSVs/4_BaroDensCompensated',
water.density = 1003,
sensor.elev = -0.02,
navd.dir = 'CSVs/5_compensated_levels_with_sensor_height'
)
}
