% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatOdyData.R
\name{clean.irregular.odyssey}
\alias{clean.irregular.odyssey}
\title{Clean Irregular Odyssey CSVs v.1}
\usage{
clean.irregular.odyssey(odyssey.file)
}
\arguments{
\item{odyssey.file}{the path to an odyssey file directly exported from the Odyssey software}
}
\description{
cleans irregular odyssey CSVs with 2 header rows, finds serial numbers from file name
outputs data frame with the following fields:
'serial.number', 'month', 'date', 'time', 'date.time', 'temperature.calibrated', 'spcond.calibrated'
}
\examples{
clean.irregular.odyssey2('PetOdy_2438.CSV')
}
