% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatOdyData.R
\name{clean.irregular.odyssey2}
\alias{clean.irregular.odyssey2}
\title{Clean Irregular Odyssey CSVs v.2}
\usage{
clean.irregular.odyssey2(odyssey.file)
}
\arguments{
\item{odyssey.file}{the path to an odyssey file directly exported from the Odyssey software}
}
\description{
cleans irregular odyssey CSVs with variable header lengths, requires a 'raw value' line
outputs data frame with the following fields:
'serial.number', 'month', 'date', 'time', 'date.time', 'temperature.calibrated', 'spcond.calibrated'
}
\examples{
clean.irregular.odyssey2('PetOdy_2438.CSV')
}
