% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatOdyData.R
\name{clean.ody.directory}
\alias{clean.ody.directory}
\title{Clean Entire Directory of Odyssey CSVs}
\usage{
clean.ody.directory(directory)
}
\arguments{
\item{directory}{directory in which data are stored}
}
\description{
runs clean.odyssey function on entire directory of raw odyssey files.
Make sure there are no other CSV files in directory.
Outputs list of data frames
}
\examples{
clean.ody.directory('~/my_odyssey_files/raw')
}
