% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatOdyData.R
\name{clean.odyssey}
\alias{clean.odyssey}
\title{Clean Odyssey CSVs}
\usage{
clean.odyssey(odyssey.file)
}
\arguments{
\item{odyssey.file}{the path to an odyssey file directly exported from the Odyssey software}
}
\description{
cleans odyssey CSVs, outputs data frame with the following fields:
'serial.number', 'month', 'date', 'time', 'date.time', 'temperature.calibrated', 'spcond.calibrated'
}
\examples{
clean.odyssey('PetOdy_2438.CSV')
}
