% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterOdy.Data.R
\name{filter_ody}
\alias{filter_ody}
\title{Filter Odyssey Data Based on Water Level Data}
\usage{
filter_ody(ody_csv, waterlevel_csv, height.offset, output.dir, ody_df,
  waterlevel_df, return.data = FALSE)
}
\arguments{
\item{ody_csv}{path to csv file of cleaned odyssey data}

\item{waterlevel_csv}{path to csv file of water-level-above-sensor data}

\item{height.offset}{height difference between sensors or threshold water level at which to accept odyssey data}

\item{output.dir}{optional directory in which to save cleaned data}

\item{ody_df}{data frame of cleaned odyssey data if ody_csv not used}

\item{waterlevel_df}{data frame of cleaned water level data if waterlevel_csv not used}

\item{return.data}{boolean, whether or not to return data as data frame}
}
\description{
Filters odyssey data to select only observations taken at the same times as a water level logger reports being inundated
Outputs optional data frame and optionally saves file
If input csv is used, output save file retains base name. Otherwise, output file is named by timestamp
}
\examples{
filter_ody(ody_csv = 'MadOdy_2437.csv', waterlevel_csv = 'Mad01-A_BaroDensCompensated.csv', height.offset = 0.07, output.dir = 'Mad01-A')
}
