% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowWaterDifference_4hrs.R
\name{get.local.lw.datums}
\alias{get.local.lw.datums}
\title{Get local low water datums}
\usage{
get.local.lw.datums(time.series, station, out.dir, by = "overall")
}
\arguments{
\item{time.series}{a csv file that contains the fields DateTime representing DateTime in excel format
and Level representing NAVD water levels in meters}

\item{station}{the numeric ID of the NOAA reference station}

\item{out.dir}{an optional output directory in which to save results}

\item{by}{whether to calculate the datums overall or by year. It is almost always most appropriate to calculate it by 'overall'.}
}
\description{
Calculates low and lower high water datums from a short term tide station using a NOAA reference station
}
\examples{
get.local.lw.datums('BrownsIsland01.csv', 9415144)
}
