% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectNOAABaro.R
\name{get.noaa.baro}
\alias{get.noaa.baro}
\title{Collect Baro Data from NOAA meteorological station}
\usage{
get.noaa.baro(station.id, start.t, end.t, interval = c("6", "h"))
}
\arguments{
\item{station.id}{a numeric station ID for the NOAA station}

\item{start.t}{the start time of the desired observations, in POSIXct format}

\item{end.t}{the end time of the desired observations, in POSIXct format}

\item{interval}{the desired interval of observation, either '6' for six minutes or 'h' for hours}
}
\description{
Collects barometric pressure data from NOAA
}
\examples{
get,noaa.baro(time1, time2, 9415115)
}
