% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcSeaWaterDensity.R
\name{seawater.density}
\alias{seawater.density}
\title{Calculate density of seawater}
\usage{
seawater.density(temperature, salinity)
}
\arguments{
\item{temperature}{temperature in degrees C}

\item{salinity}{salinity in PSU}
}
\description{
Calculates density in kg/m^3, after Millero and Poisson (1981)
}
\examples{
seawater.density(25, 30)
}
