% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardizeLoggerData.R
\name{standardize.baro.folder}
\alias{standardize.baro.folder}
\title{Standardize Entire Folder of Solinst Barologger Data}
\usage{
standardize.baro.folder(baro.dir)
}
\arguments{
\item{baro.dir}{the path to the whole directory of solinst csv files}
}
\description{
This function cleans a whole folder of barometric pressure data exported from the Levelogger software
}
\examples{
standardize.baro.folder('myproject/raw_baro_data')
}
