% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardizeLoggerData.R
\name{standardize.baro}
\alias{standardize.baro}
\title{Standardize Solinst Barologger Data}
\usage{
standardize.baro(solinst.file)
}
\arguments{
\item{solinst.file}{the path to your solinst csv file}
}
\description{
This function cleans barometric pressure data exported from the Levelogger software
}
\examples{
standardize.baro('HumboldtBaro_120815.csv')
}
