% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardizeLoggerData.R
\name{standardize.hobo}
\alias{standardize.hobo}
\title{Standardize Hobo Levelogger Data}
\usage{
standardize.hobo(hobo.file, logger.id = NULL, temperature = NULL)
}
\arguments{
\item{hobo.file}{the path to your solinst csv file}

\item{logger.id}{the serial number of your logger, necessary for temperature compensation}

\item{temperature}{data frame of an external source of temperature data,
if the logger was launched without temperature. Must have fields 'time' in POSIX format
and 'temperature' in degrees C}
}
\description{
This function cleans water level data exported from the Hoboware software
}
\examples{
standardize.hobo('Petaluma01_120815.csv')
standardize.hobo('Petaluma01_110412.csv', logger.id = 11201228, temperature = my.temps)
}
