% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardizeLoggerData.R
\name{standardize.solinst}
\alias{standardize.solinst}
\title{Standardize Solinst Levelogger Data}
\usage{
standardize.solinst(solinst.file, programmed.h2o.density = 1000)
}
\arguments{
\item{solinst.file}{the path to your solinst csv file}

\item{programmed.h20.density}{the water density programmed into the logger on deployment, in g/L. Defaults to 1000}
}
\description{
This function converts water level data exported from the Levelogger software into kilopascals
}
\examples{
standardize.solinst('Petaluma01_120815.csv')
}
