% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.aware.tp2wlas.R
\name{tp.to.wlas.vardens}
\alias{tp.to.wlas.vardens}
\title{Calculate compensated water levels above sensor using monthly water densities}
\usage{
tp.to.wlas.vardens(logger.data, baro.data, wq.df)
}
\arguments{
\item{logger.data}{a data frame of standardized logger pressure data. Must have fields 'time', 'total.pressure'}

\item{baro.data}{a data frame of baro pressure data. Must have fields 'time', 'baro.pressure'}

\item{wq.df}{is a data frame of water quality by month and year. Must have month and year as
'Mo' and 'Yr' respectively and mean temperature and salinity as 'MeanT' and 'MeanS' respectively}
}
\description{
This function performs barometric compensation and density compensation using monthly water quality data
}
\examples{
tp.to.wlas(petaluma.standardized, bj.baro.standard, 1080.34)
}
