% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighWaterDifference_4hrs.R
\name{wll.data.stats}
\alias{wll.data.stats}
\title{Get water level data stats}
\usage{
wll.data.stats(wll.data.path)
}
\arguments{
\item{wll.data.path}{a csv file that contains the fields DateTime representing DateTime in excel format
and Level representing NAVD water levels in meters}
}
\description{
calculate start and end dates, highest observed water levels, gaps, etc.
}
\examples{
wll.data.stats('Petaluma.csv')
}
